Move admin auth to global auth system
