+++ packages/core/admin/server/routes/permissions.js
@@ -1,0 +1,20 @@
+'use strict';
+
+module.exports = [
+  {
+    method: 'GET',
+    path: '/permissions',
+    handler: 'permission.getAll',
+    config: {
+      policies: ['admin::isAuthenticatedAdmin'],
+    },
+  },
+  {
+    method: 'POST',
+    path: '/permissions/check',
+    handler: 'permission.check',
+    config: {
+      policies: ['admin::isAuthenticatedAdmin'],
+    },
+  },
+];
