+++ packages/core/admin/server/routes/roles.js
@@ -1,0 +1,67 @@
+'use strict';
+
+module.exports = [
+  {
+    method: 'POST',
+    path: '/users/batch-delete',
+    handler: 'user.deleteMany',
+    config: {
+      policies: [{ name: 'admin::hasPermissions', options: { actions: ['admin::users.delete'] } }],
+    },
+  },
+  {
+    method: 'GET',
+    path: '/roles/:id/permissions',
+    handler: 'role.getPermissions',
+    config: {
+      policies: [
+        'admin::isAuthenticatedAdmin',
+        { name: 'admin::hasPermissions', options: { actions: ['admin::roles.read'] } },
+      ],
+    },
+  },
+  {
+    method: 'PUT',
+    path: '/roles/:id/permissions',
+    handler: 'role.updatePermissions',
+    config: {
+      policies: [
+        'admin::isAuthenticatedAdmin',
+        { name: 'admin::hasPermissions', options: { actions: ['admin::roles.update'] } },
+      ],
+    },
+  },
+  {
+    method: 'GET',
+    path: '/roles/:id',
+    handler: 'role.findOne',
+    config: {
+      policies: [
+        'admin::isAuthenticatedAdmin',
+        { name: 'admin::hasPermissions', options: { actions: ['admin::roles.read'] } },
+      ],
+    },
+  },
+  {
+    method: 'GET',
+    path: '/roles',
+    handler: 'role.findAll',
+    config: {
+      policies: [
+        'admin::isAuthenticatedAdmin',
+        { name: 'admin::hasPermissions', options: { actions: ['admin::roles.read'] } },
+      ],
+    },
+  },
+  {
+    method: 'PUT',
+    path: '/roles/:id',
+    handler: 'role.update',
+    config: {
+      policies: [
+        'admin::isAuthenticatedAdmin',
+        { name: 'admin::hasPermissions', options: { actions: ['admin::roles.update'] } },
+      ],
+    },
+  },
+];
