+++ packages/core/strapi/lib/services/server/admin-api.js
@@ -5,6 +5,7 @@ const { createAPI } = require('./api');
 const createAdminAPI = strapi => {
   const opts = {
     prefix: '', // '/admin';
+    type: 'admin',
   };
 
   return createAPI(strapi, opts);
