+++ packages/core/strapi/lib/services/server/policy.js
@@ -7,10 +7,14 @@ const { bodyPolicy } = policy;
 
 const getPoliciesConfig = propOr([], 'config.policies');
 
-const resolvePolicies = (route, opts = {}) => {
+const resolvePolicies = route => {
+  const { pluginName, apiName } = route.info || {};
   const policiesConfig = getPoliciesConfig(route);
 
-  const policies = policiesConfig.map(policyName => policy.get(policyName, opts));
+  const policies = policiesConfig.map(policyName => {
+    return policy.get(policyName, { pluginName, apiName });
+  });
+
   return [...policies, bodyPolicy];
 };
