+++ packages/plugins/users-permissions/server/register.js
@@ -3,5 +3,5 @@
 const authStrategy = require('./auth/strategy');
 
 module.exports = strapi => {
-  strapi.container.get('content-api').auth.register(authStrategy);
+  strapi.container.get('auth').register('content-api', authStrategy);
 };
